dashboard/requirements.txt
streamlit==1.28.0
pandas==2.1.1
plotly==5.17.0
requests==2.31.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1